+++ src/Contracts/SaveSingleImageHandlerInterface.php
@@ -4,5 +4,5 @@ namespace DevMahmoudMustafa\ImageKit\Contracts;
 
 interface SaveSingleImageHandlerInterface extends ImageHandlerInterface
 {
-    public function saveImage(): string;
+    public function saveImage(): string|array;
 }